+++ app/controllers/players.rb
@@ -7,13 +7,15 @@ Observatory::App.controllers :players do
 
     search_param = params['filter']
 
-    begin
-      last_active_after = Date.strptime(
-        params.fetch('last_active_after', ''),
-        '%Y-%m-%d'
-      )
-    rescue ArgumentError
-      last_active_after = nil
+    last_active_after = nil
+    if params.key? 'last_active_after'
+      begin
+        last_active_after = Date.strptime(
+          params.fetch('last_active_after'),
+          '%Y-%m-%d'
+        )
+      rescue ArgumentError
+      end
     end
 
     direct_results = []
@@ -82,7 +84,10 @@ Observatory::App.controllers :players do
       }
 
       direct_results = direct_results.select do |player|
-        player.last_activity >= last_active_after
+        # last_activity might be nil if the player has just been added, in
+        # which case a direct hit (by Steam ID) will be possible, but no data
+        # is on record yet.
+        player.last_activity >= last_active_after if player.last_activity
       end
     else
       logger.debug 'Skipping filtering by last activity.'