+++ CHANGELOG.md
@@ -8,9 +8,13 @@ glance - what to expect from upgrading to a new version.
 
 ### Added
 
+- PAX2012 badge, which seems to be exported by Hive now.
+
 ### Changed
 
 ### Fixed
+
+- Exception if unknown badge encountered in Hive response.
 
 ### Security